+++ src/Observers/AttributeDefaultOptionObserver.php
@@ -1,0 +1,133 @@
+<?php
+
+/**
+ * TechDivision\Import\Attribute\Observers\AttributeDefaultOptionObserver
+ *
+ * NOTICE OF LICENSE
+ *
+ * This source file is subject to the Open Software License (OSL 3.0)
+ * that is available through the world-wide-web at this URL:
+ * http://opensource.org/licenses/osl-3.0.php
+ *
+ * PHP version 5
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+
+namespace TechDivision\Import\Attribute\Observers;
+
+use TechDivision\Import\Attribute\Utils\ColumnKeys;
+use TechDivision\Import\Attribute\Utils\MemberNames;
+use TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface;
+use TechDivision\Import\Utils\StoreViewCodes;
+
+/**
+ * Observer that updates the attribute with the default value found in the CSV file.
+ *
+ * @author    Tim Wagner <t.wagner@techdivision.com>
+ * @copyright 2016 TechDivision GmbH <info@techdivision.com>
+ * @license   http://opensource.org/licenses/osl-3.0.php Open Software License (OSL 3.0)
+ * @link      https://github.com/techdivision/import-attribute
+ * @link      http://www.techdivision.com
+ */
+class AttributeDefaultOptionObserver extends AbstractAttributeImportObserver
+{
+
+    /**
+     * The attribute processor instance.
+     *
+     * @var \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface
+     */
+    protected $attributeBunchProcessor;
+
+    /**
+     * Initializes the observer with the passed subject instance.
+     *
+     * @param \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface $attributeBunchProcessor The attribute bunch processor instance
+     */
+    public function __construct(AttributeBunchProcessorInterface $attributeBunchProcessor)
+    {
+        $this->attributeBunchProcessor = $attributeBunchProcessor;
+    }
+
+    /**
+     * Process the observer's business logic.
+     *
+     * @return void
+     */
+    protected function process()
+    {
+
+        // load the default value we want to set
+        $defaultValue = $this->getValue(ColumnKeys::DEFAULT_VALUE);
+
+        // load the entity type ID for the value from the system configuration
+        $entityTypeId = $this->getEntityTypeId();
+
+        // initialize the data to load the EAV attribute option
+        $storeId = $this->getRowStoreId(StoreViewCodes::ADMIN);
+        $attributeCode = $this->getValue(ColumnKeys::ATTRIBUTE_CODE);
+
+        // try to load the EAV attribute option
+        if ($attributeOption = $this->loadAttributeOptionByEntityTypeIdAndAttributeCodeAndStoreIdAndValue($entityTypeId, $attributeCode, $storeId, $defaultValue)) {
+            // load the EAV attribute with the actual code
+            $attribute = $this->loadAttributeByEntityTypeIdAndAttributeCode($entityTypeId, $attributeCode);
+            // set the default value with the EAV attribute option ID and update the attribute
+            $this->persistAttribute($this->mergeEntity($attribute, array(MemberNames::DEFAULT_VALUE => $attributeOption[MemberNames::OPTION_ID])));
+        }
+    }
+
+    /**
+     * Return's the attribute bunch processor instance.
+     *
+     * @return \TechDivision\Import\Attribute\Services\AttributeBunchProcessorInterface The attribute bunch processor instance
+     */
+    protected function getAttributeBunchProcessor()
+    {
+        return $this->attributeBunchProcessor;
+    }
+
+    /**
+     * Persist the passed attribute.
+     *
+     * @param array $attribute The attribute to persist
+     *
+     * @return void
+     */
+    protected function persistAttribute(array $attribute)
+    {
+        return $this->getAttributeBunchProcessor()->persistAttribute($attribute);
+    }
+
+    /**
+     * Return's the EAV attribute with the passed entity type ID and code.
+     *
+     * @param integer $entityTypeId  The entity type ID of the EAV attribute to return
+     * @param string  $attributeCode The code of the EAV attribute to return
+     *
+     * @return array The EAV attribute
+     */
+    protected function loadAttributeByEntityTypeIdAndAttributeCode($entityTypeId, $attributeCode)
+    {
+        return $this->getAttributeBunchProcessor()->loadAttributeByEntityTypeIdAndAttributeCode($entityTypeId, $attributeCode);
+    }
+
+    /**
+     * Load's and return's the EAV attribute option with the passed entity type ID, code, store ID and value.
+     *
+     * @param string  $entityTypeId  The entity type ID of the EAV attribute to load the option for
+     * @param string  $attributeCode The code of the EAV attribute option to load
+     * @param integer $storeId       The store ID of the attribute option to load
+     * @param string  $value         The value of the attribute option to load
+     *
+     * @return array The EAV attribute option
+     */
+    protected function loadAttributeOptionByEntityTypeIdAndAttributeCodeAndStoreIdAndValue($entityTypeId, $attributeCode, $storeId, $value)
+    {
+        return $this->getAttributeBunchProcessor()->loadAttributeOptionByEntityTypeIdAndAttributeCodeAndStoreIdAndValue($entityTypeId, $attributeCode, $storeId, $value);
+    }
+}